Reorder task module code
