+++ src/wazuh_modules/task_manager/wm_task_manager.h
@@ -25,15 +25,6 @@ typedef struct _wm_task_manager {
     int cleanup_time;
 } wm_task_manager;
 
-typedef enum _upgrade_status {
-    WM_TASK_UPGRADE_ERROR = 0,
-    WM_TASK_UPGRADE_UPDATING,
-    WM_TASK_UPGRADE_UPDATED,
-    WM_TASK_UPGRADE_OUTDATED,
-    WM_TASK_UPGRADE_TIMEOUT,
-    WM_TASK_UPGRADE_LEGACY
-} upgrade_status;
-
 typedef enum _error_code {
     WM_TASK_SUCCESS = 0,
     WM_TASK_INVALID_MESSAGE,
@@ -49,8 +40,6 @@ typedef enum _error_code {
 
 extern const wm_context WM_TASK_MANAGER_CONTEXT;   // Context
 
-extern const char* task_statuses[];
-
 // Parse XML configuration
 int wm_task_manager_read(xml_node **nodes, wmodule *module);
 
@@ -60,14 +49,7 @@ int wm_task_manager_read(xml_node **nodes, wmodule *module);
  * @param response Response to be sent to the connection.
  * @return Size of the response string.
  * */
-size_t wm_task_manager_dispatch(const char *msg, char **response);
-
-/**
- * Parse the incomming message and return a JSON with the message.
- * @param msg Incomming message from a connection.
- * @return JSON array when succeed, NULL otherwise.
- * */
-cJSON* wm_task_manager_parse_message(const char *msg);
+size_t wm_task_manager_dispatch(const char *msg, char **response) __attribute__((nonnull));
 
 /**
  * Analyze a task by module and call appropiate analyze function.
@@ -75,7 +57,7 @@ cJSON* wm_task_manager_parse_message(const char *msg);
  * @param error_code Variable to store an error code if something is wrong.
  * @return JSON object with the response for this task.
  * */
-cJSON* wm_task_manager_analyze_task(const cJSON *task_object, int *error_code);
+cJSON* wm_task_manager_analyze_task(const cJSON *task_object, int *error_code) __attribute__((nonnull));
 
 /**
  * Analyze a api task command.
@@ -85,7 +67,7 @@ cJSON* wm_task_manager_analyze_task(const cJSON *task_object, int *error_code);
  * @param task_id Task id extracted from task_object. 
  * @return JSON object with the response for this task.
  * */
-cJSON* wm_task_manager_analyze_task_api_module(const cJSON *task_object, int *error_code, int agent_id, int task_id);
+cJSON* wm_task_manager_analyze_task_api_module(const cJSON *task_object, int *error_code, int agent_id, int task_id) __attribute__((nonnull));
 
 /**
  * Analyze a upgrade_module task by command. Update the tasks DB when necessary.
@@ -96,93 +78,7 @@ cJSON* wm_task_manager_analyze_task_api_module(const cJSON *task_object, int *er
  * @param status Status extracted from task_object.
  * @return JSON object with the response for this task.
  * */
-cJSON* wm_task_manager_analyze_task_upgrade_module(const cJSON *task_object, int *error_code, int agent_id, int task_id, char *status);
-
-/**
- * Build a JSON object response.
- * 
- * Example 1: Success
- * 
- * {
- *     "error":0,
- *     "data":"Success",
- *     "agent":4,
- *     "task_id":2
- * }
- * 
- * Example 2: Invalid message
- * 
- * {
- *     "error":1,
- *     "data":"Invalid message"
- * }
- * 
- * 
- * Example 3: Invalid module
- * 
- * {
- *     "error":2,
- *     "data":"Invalid module",
- *     "task_id":45
- * }
- * 
- * Example 4: Invalid command
- * 
- * {
- *     "error":3,
- *     "data":"Invalid command",
- *     "agent":15
- * }
- * 
- * @param error_code Code of the error.
- * @param agent_id ID of the agent when receiving a request for a specific agent.
- * @param task_id ID of the task when receiving a request for a specific task.
- * @param status Status of the task when receiving a request for a specific status.
- * @return JSON object.
- * */
-cJSON* wm_task_manager_build_response(int error_code, int agent_id, int task_id, char *status);
-
-/**
- * Add data to a JSON object response.
- * 
- * Example 1: On upgrade_result request
- * 
- * {
- *     "error":0,
- *     "data":"Success",
- *     "agent":4,
- *     "task_id":2,
- *     "module":"upgrade_module",
- *     "command":"upgrade",
- *     "status":"Updating"
- *     "create_time":"2020-08-06 22:51:44"
- *     "update_time":"2020-08-06 22:53:21"
- * }
- * 
- * Example 1: On task_result request
- * 
- * {
- *     "error":0,
- *     "data":"Success",
- *     "agent":4,
- *     "task_id":2,
- *     "module":"upgrade_module",
- *     "command":"upgrade",
- *     "status":"In progress"
- *     "create_time":"2020-08-06 22:51:44"
- *     "update_time":"2020-08-06 22:53:21"
- * }
- * 
- * @param response JSON object response
- * @param module Module of the task when receiving a request for a specific task.
- * @param command Command of the task when receiving a request for a specific task.
- * @param status Status of the task when receiving a request for a specific task.
- * @param create_time Date of creation task.
- * @param last_update_time Date of update task.
- * @param request_command Command that requested the query.
- * @return JSON object.
- * */
-cJSON* wm_task_manager_build_response_result(cJSON *response, const char *module, const char *command, char *status, int create_time, int last_update_time, char *request_command);
+cJSON* wm_task_manager_analyze_task_upgrade_module(const cJSON *task_object, int *error_code, int agent_id, int task_id, char *status) __attribute__((nonnull));
 
 #endif
 #endif
